/booking - Logout fixes
Could have trouble clearing the session with gift cards in the cart

+++ booking/laravel/app/controllers/LogoutController.php
@@ -15,8 +15,11 @@ class LogoutController extends BaseController
         {
             foreach ($cart as $cartItemId => $cartItem)
             {
-                $onlineBookingsReservationId = Cart::getOnlineBookingsReservationId($cartItemId);
-                CS_API::deleteOnlineReservation($onlineBookingsReservationId); //Remove the online booking
+                if ($cart[$cartItemId]['type'] == 'heat')
+                {
+                    $onlineBookingsReservationId = Cart::getOnlineBookingsReservationId($cartItemId);
+                    CS_API::deleteOnlineReservation($onlineBookingsReservationId); //Remove the online booking
+                }
                 Cart::removeFromCart($cartItemId); //Then remove it from the local cart
             }
         }
